% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_curve_slope.R
\name{avg_curve_slope}
\alias{avg_curve_slope}
\title{Get average curve slopes}
\usage{
avg_curve_slope(Dataframe, Trial)
}
\arguments{
\item{Dataframe}{a GCalcium-format data frame or matrix}

\item{Trial}{a single trial number}
}
\value{
vector of average rate of change for each curve
}
\description{
`avg_curve_slope` finds the average curve slopes for a trial using inflect_points_df and lm
}
\note{
curves of a single value will result in the average slope being labeled as NA
}
\examples{
df.new <- format_data(data = GCaMP)
avg_curve_slope(Dataframe = df.new, Trial = 1)
}
