% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_trials.R
\name{center_trials}
\alias{center_trials}
\title{Centers individual or multiple trials}
\usage{
center_trials(Dataframe, Trials, Baseline.times, Baseline.frame = NULL,
  Method = "perc.baseline", Summary.type = "mean")
}
\arguments{
\item{Dataframe}{a GCalcium-format data frame or matrix}

\item{Trials}{a single trial number or vector of trial numbers to center}

\item{Baseline.times}{range of time from Baseline.frame to compute the baseline value from}

\item{Baseline.frame}{a GCalcium-format data frame or matrix containing the baseline period. If frame is not specified, Dataframe is automatically used}

\item{Method}{the type of centering to be used}

\item{Summary.type}{the type of summary statistic to be centered by}
}
\value{
a GCalcium-format data frame with centered values
}
\description{
`center_trials` subtracts summarized data during a user-specified time period from trial values
}
\examples{
### Format data frame
df.new <- format_data(GCaMP)

### Transform into percent baseline: relative to -3s to -1s before epoc
center_trials(Dataframe = df.new, Trials = c(1, 2, 3), Baseline.times = c(-3, -1))
}
