% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Reformat and label data into GCalcium format}
\usage{
format_data(data)
}
\arguments{
\item{data}{A data set with observation times in the first row or column, and observed values from trials in each following row or column.}
}
\value{
Data frame with labeled time and trial columns
}
\description{
`format_data` changes the format of data from wide time series to long format with labeled columns. If data is already in long format, the function simply re-labels the columns
}
\examples{
df.new <- format_data(GCaMP)
}
