% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCaMP-data.R
\docType{data}
\name{GCaMP}
\alias{GCaMP}
\title{Pre-filtered GCaMP calcium activity waveforms}
\format{A data frame with 11 rows and 814 columns
\describe{
  \item{Time}{observed times of Trial.n values, in seconds}
  \item{Trial.n}{values of calcium activity signals for trial n}
  ...
}}
\source{
Dr. David Root, Maier-Watkins laboratory
}
\usage{
data(GCaMP)
}
\description{
A dataset containing calcium activity and
observed times during a behavioral paradigm pilot study;
organized into 10 individual trials
}
\keyword{datasets}
