% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect_points.R
\name{inflect_points}
\alias{inflect_points}
\title{Find inflection pooints}
\usage{
inflect_points(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
Inflection points of the input vector
}
\description{
`inflect_points` finds inflection points of waveform data
}
\examples{
df.new <- format_data(GCaMP)
inflect_points(df.new$Trial1)
}
