% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect_points_df.R
\name{inflect_points_df}
\alias{inflect_points_df}
\title{Find inflection points and curve numbers}
\usage{
inflect_points_df(Dataframe, Trial)
}
\arguments{
\item{Dataframe}{a GCalcium-format data frame or matrix}

\item{Trial}{a single trial number}
}
\value{
data frame with variables indicating the time, raw values, curve numbers, and inflection points corresponding to each data point of the input
}
\description{
`inflect_points_df` finds inflection points of activity waves, returning a summarized data frame
}
\examples{
df.new <- format_data(GCaMP)
inflect_points_df(Dataframe = df.new, Trial = 1)
}
