% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_baseline.R
\name{perc_baseline}
\alias{perc_baseline}
\title{Find the percent change from baseline}
\usage{
perc_baseline(xvals, base.val)
}
\arguments{
\item{xvals}{vector of numbers}

\item{base.val}{the baseline value to be centered by}
}
\value{
a vector with values transformed to percent baseline
}
\description{
`perc_baseline` calculates the percent change from a user-specified baseline
}
\examples{
### Format data frame
# Format data
df.new <- format_data(GCaMP)

# Split data
basevals <- df.new$Trial1[df.new$Time <= 0]
eventvals <- df.new$Trial1[df.new$Time > 0]

# Find baseline (pre-epoc) values
base.mean <- mean(basevals)

# Compute values
perc_baseline(x = eventvals, base.val = base.mean)
}
